Fixed realm crashing when updating alerts and exiting map
Added a delete tag to highway alerts items so they are not removed
while app is running. The crash happened when an update was started on
the traffic map and user navigates back to home screen where a highest
alert was deleted while being used by the ticker view.

Updated all get methods that use delete tag to only return data with
delete = false