Fixes #96
By returning true every time in
splitViewController(_:collapseSecondary:onto:) the primary view
(in this case home view) always covered the detail view (in this case
the MapView, or any detail view in the app).

To fix this, the homeViewController now sets a shouldCollapse boolean
in HomeSplitViewController using the setShouldCollapseFalse() accessor
method.