Refactored SplitViewController
iOS 9 lets us segue to storyboard refs. This means we can move most of
the navigation logic in the HomeViewController to the story board.
Added highlighting for the current selected view while in split mode.
Tapping the current selected section now returns user to the top of the
navigation stack instead of making a new segue.