Fixes go to location issue, adds Everett to list
Adds Everett to the GoTo Locations list.

Fixes issue introduced in commit 204cfa11 where the go to locations
menu fails to work on iPad or devices that display the menu as a pop up.

The menu had previously only set the location defaults, which are never
read read when the menu is a pop up.