Ferry timestamp fixes
The ferry alerts section wasn’t set up to handle the new timestamp
format, which caused a crash when loading it. This has been fixed.

The day picker will now also generate days of the week based on how
many are returned from the API and not just a set 7